Changed password has to use the PBKDF2 method.  This is NOT COMPATIBLE with previous implementation.  This also changes table structures over previous versions.
Created an experimental install controller that creates the tables.  THis is -very- experimental.  Eventually I would like to have it build the config file for you to simplify installation.

Added a new abstraction layer library in logrus/Password.php .  This is a connector to allow you to abstract out the management of the member account and passwords.  This allows you to more easily separate out the hash generation.  For example offloading it to a remote server instead of storing/managing locally.